@@ -1,6 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-import Background from './Background';
+
+import { Background } from './Background';
 
 test('renders learn react link', () => {
 	render(<Background />);
